20DEC

+++ Fashion/Stores/FashionStore.m
@@ -27,7 +27,7 @@
 #pragma mark news requests
 
 - (void)requestNews:(NSInteger)page withCompletion:(void(^)(NSArray *news, NSError *error))completion {
-    NSURL *url = [NSURL URLWithString:[NSString stringWithFormat:@"http://fashion.ie/api/get_posts?page=%lu",page]]; // Construct URL
+    NSURL *url = [NSURL URLWithString:[NSString stringWithFormat:@"http://fashion.ie/api/get_posts?page=%lu",(long)page]]; // Construct URL
     [NSURLSession jsonFromURL:url completion:^(id json){
         NSArray *news = [self isListJsonOK:json] ? [self newsArrayWithJSON:json[KEY_RESULTS]] : nil; // Get the result
         dispatch_async_main(^{
